# openapi.yml
openapi: "3.0.2"
info: 
  title: Simple Login
  version: "1.0"
  contact: 
    email: "wilman262@gmail.com"

servers:
  - url: http://localhost:3000/api/v1

components:
  schemas:
    User:
      allOf: 
        - $ref: "#/components/schemas/NewUser"
        - type: object
          properties:
            followers:
              type: number
            following:
              type: array
              items:
                type: string
            _id:
              type: string
            preferences:
              items:
                type: string
              type: array
            registrationType:
              enum:
                - local
                - Facebook
                - Twitter
              type: string

    NewUser:
      properties:
        email:
          type: string
        facebook:
          type: string
        name:
          type: string
        password:
          type: string
        perfilImage:
          type: string
        surname:
          type: string
        username:
          type: string
      type: object
      required:
        - name
        - email
        - password
        - username

    AuthResponse:
      additionalProperties: false
      properties:
        data:
          type: string
        message:
          type: string
        status:
          type: integer
        id:
          type: string
      type: object

    Entity:
      additionalProperties: false
      properties:
        data:
          type: object
        status:
          type: integer
      type: object

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


paths:
  /auth/login:
    post:
      security:
        - {}
      summary: Creates a bearer token for a user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        "200":
          description: Genereted jwt token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/signup:
    post:
      security:
        - {}
      summary: Creates a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUser"
            example:
              name : Wilman
              email: "wilman262@gmail.com"
              username: Wilch
              password: "123456789"
      responses:
        "201":
          description: Genereted jwt token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /user:
    get:
      description: Obtiene un usuario
      tags: [User]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf: 
                  - $ref: "#/components/schemas/Entity"
                  - type: object
                    properties:
                      data:
                        type: object
                        $ref: "#/components/schemas/User"
security:
  - bearerAuth: ["/auth/login"]
  - {}  